pyproject.toml
[tool.poetry]
name = "when2meet"
version = "0.1.0"
description = ""
authors = ["박준혁 <jun201908@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
orjson = "3.10.12"
cryptography = "44.0.0"
aerich = "0.8.1"
tomlkit = "0.13.2"
pydantic-settings = "2.7.1"
mock = "^5.2.0"
patch = "^1.16"
httpx = "0.28.1"

[tool.poetry.dependencies.tortoise-orm]
version = "0.23.0"
extras = ["asyncmy"]

[tool.poetry.group.dev.dependencies]
black = "24.10.0"
ruff = "0.8.2"
mypy = "1.13.0"
pytest = "8.3.4"
coverage = "7.6.9"
pytest-asyncio = "0.25.0"

[tool.black]
line-length = 120

[tool.ruff]
target-version = "py313"

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = 3.13
strict = true

[tool.coverage.run]
omit = ["*/test_*.py"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
filterwarnings = ['ignore::DeprecationWarning']

[tool.aerich]
tortoise_orm = "app.configs.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
